Module type Bits.Full30

type state

include State30 with type state := state

type state

The type of PRNG states.

val bits : state -> int

Provide 30 random bits in an integer.

include Init30 with type state := state

type state

val new_state : unit -> state

Create a new non-initialised state.

val assign : state -> state -> unit

assign s1 s2 copies s2 into s1.

val init_size : int

Required size for the array in init.

val init : state -> int array -> unit

Initialise the state based on the values in the array. The array contains init_size random 30-bits integers.

val default_seed : int

Seed that will be given to init to generate the default state.