Module type Bits.FULLHI64

type state

val bits : state -> int64

val new_state : unit -> state

val assign : state -> state -> unit

val full_init_size : int

Number of int64 numbers required by full_init.

val full_init : state -> int64 array -> unit

Similar to FULL.full_init except it receives an array of size full_init_size containing “good” int64 values.

val default_seed : int

Default seed instead of default state. The state will be obtained by calling new_state and then full_init on an array derived from the seed, similar to what Random does.