Module MakeRandom__.Full30Init64

Full30Init64

This module contains one unique functor which, provided a type state, a function bits that generates bits from the state and a few functions to initialise and manipulate the state, returns a module with the same signature as Stdlib.Random. It is basically only a convenience on top of State30 and Init30.

This is the 30/64-bits version: the function bits is expected to return an int whose 30 lower bits only are set, and the function init is expected to take an array of int64 whose 64 bits are set. This version is meant to be used for PRNGs whose state is made of 64-bits integers and who generate 30-bits outputs (eg. the bindings for xoshiro256++).

val random_seed : unit -> int array

module Make : functor (B : MakeRandom.Bits.Full30Init64) -> MakeRandom.Sig.Full