Module MakeRandom__.Init30

Init30

This module contains one unique functor which, provided a type state and a few functions to initialise and manipulate it, returns a module similar to that of Stdlib.Random.State, containing all the functions that initialise and manipulate the state.

In particular, this functor provide convenient initialisation: it takes a number init_size and a function init able to initialise a state when given an array of size init_size containing already "good" bits and derives from it the function full_init generating "good" bits out of an array of anything containing any kind integers (the array can be empty and can contain only low-entropy integers).

This is the 30-bits version: the function init takes an array of int whose 30 lower bits only are set. It is meant to be used for PRNGs whose state is made of 30-bits integers (eg. the one from the standard library).

val random_seed : unit -> int array

module Make : functor (B : MakeRandom.Bits.Init30) -> sig ... end