Module MakeRandom__.State30

State30

This module contains one unique functor which, provided a type state and a function bits generates bits from the state, returns a module similar to that of Stdlib.Random.State, containing all the functions that take the state and generate bits from it.

This is the 30-bits version: the function bits is expected to return an int whose 30 lower bits only are set.

module Make : functor (B : MakeRandom.Bits.State30) -> sig ... end