Module MakeRandom__Bits

Bits Providers

This module contains the signatures of the modules that can be given to the various functors in MakeRandom, specifying how to provide bits. BASIC only provides bits, while FULL provides bits as well as state manipulations.

Basic

To use in MakeRandom.Basic.

module type Basic30 = sig ... end

module type Basic64 = sig ... end

State

module type State30 = sig ... end

module type State64 = sig ... end

Init

module type Init30 = sig ... end

module type Init64 = sig ... end

Full

module type Full30 = sig ... end

module type Full64 = sig ... end

module type Full30Init64 = sig ... end

module type Full64Init30 = sig ... end