Module MakeRandom__Sig

Signatures

This module contains the signatures of the modules generated by functors in MakeRandom. BASIC contains only the basic/usual functions from the Random interface, while FULL also contains the Random.State module and the state manipulations functions.

Type of the basic functions

The module type BASIC includes the basic/usual functions from the Random interface. For initialisation functions and PRNG state, see FULL.

module type BASIC = sig ... end

Complete type of the module

The module type FULL is identical to that of Random except for the type FULL.State.t which is, of course, different. It can be used as a drop-in replacement of Random.

module type FULL = sig ... end