Module MakeRandom.Bits

Signatures for providers of bits (input signatures of the functors in this module).

Bits Providers

This module contains the signatures of the modules that can be given to the various functors in MakeRandom, specifying how to provide bits. BASIC only provides bits, while FULL provides bits as well as state manipulations.

Basic

To use in MakeRandom.Basic.

module type BASIC = sig ... end

module type BASIC64 = sig ... end

Full

To use in MakeRandom.Full.

module type FULL = sig ... end

module type FULL64 = sig ... end

Full with Helper for Initialisation

module type FULLHI64 = sig ... end